+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - [About] PH language support
 - [Twitch] Embedded player
+- [Twitch] History
 
 ## [0.2.4] - 2024-08-28
 ### Added